+++ modules/invenio-iiif/CHANGES.rst
@@ -1,6 +1,6 @@
 ..
     This file is part of Invenio.
-    Copyright (C) 2018 CERN.
+    Copyright (C) 2018-2020 CERN.
 
     Invenio is free software; you can redistribute it and/or modify it
     under the terms of the MIT License; see LICENSE file for more details.
@@ -8,6 +8,19 @@
 Changes
 =======
 
-Version 1.0.0a4 (released 2018-09-11)
+Version 1.2.0 (released 2021-10-20)
+
+- Unpin Flask.
+
+Version 1.1.1 (released 2021-05-27)
+
+- Make API decorator and image opener handler configurable.
+
+Version 1.1.0 (released 2020-03-19)
+
+- Remove Python 2.7 support.
+- Bump Flask-IIIF to v0.6.
+
+Version 1.0.0 (released 2019-07-24)
 
 - Initial public release.